+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * resolved modal issue, integrated api for update property ([56fc7b3](https://gitlab.com/paras205/boilerplate-react/commit/56fc7b3dce0062608340685fa86dcee7ab0d45da))
 * table changes ([efc1491](https://gitlab.com/paras205/boilerplate-react/commit/efc1491a997cc0f44fd7720a21fed789d69c01df))
 * table changes ([b545c66](https://gitlab.com/paras205/boilerplate-react/commit/b545c668a87aa21a950ccbfd387d5cc97bd70ad4))
+* voucher changes ([e32131c](https://gitlab.com/paras205/boilerplate-react/commit/e32131c325d7e2e1b7083424ae78b39822a68658))